fix entity name
scene table을 추가함으로써 클래스 간의 관계, 이름이 변경됨. 이 점을 수정
